rooms page

+++ admin/ajax/features_facilities.php
@@ -40,9 +40,17 @@
         $frm_data = filteration($_POST);
         $values=[$frm_data['rem_feature']];
 
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q,$values,'i');
-        echo $res;
+        $check_q=select('SELECT * FROM `room_features` WHERE `features_id`=?',[$frm_data['rem_feature']],'i');
+
+        if(mysqli_num_rows($check_q)==0){ 
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q,$values,'i');
+            echo $res;
+        }
+        else {
+            echo 'room_added';
+        }
+
     }
 
     if(isset($_POST['add_facility'])){
@@ -100,9 +108,27 @@
         $frm_data = filteration($_POST);
         $values=[$frm_data['rem_facility']];
 
-        $q = "DELETE FROM `facilities` WHERE `id`=?";
-        $res = delete($q,$values,'i');
-        echo $res;
+        $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
+        $res = select($pre_q,$values,'i');
+        $img = mysqli_fetch_assoc($res);
+        
+        $check_q=select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?',[$frm_data['rem_facility']],'i');
+
+        if(mysqli_num_rows($check_q)==0){ 
+            if(deleteImage($img['icon'],FEATURES_FOLDER)){ 
+                $q = "DELETE FROM `facilities` WHERE `id`=?";
+                $res = delete($q,$values,'i');
+                echo $res;
+            }
+            else{
+                echo 0;
+            }
+        }
+        else{
+            echo 'room_added';
+        }
+
+
     }
     
 ?>
